Create player.rb file
