+++ bin/main.rb
@@ -1,6 +1,7 @@
 #!/usr/bin/env ruby
 require './lib/game'
 require './lib/board'
+require './lib/player'
 require './lib/clear'
 
 game_board = Board.new
