+++ lib/game.rb
@@ -1,14 +1,5 @@
 require 'set'
 
-class Player
-  attr_reader :name
-  attr_accessor :choices
-  def initialize(name)
-    @name = name
-    @choices = []
-  end
-end
-
 class Rules
   def initialize
     @victory_rows = [%w[1 2 3], %w[4 5 6], %w[7 8 9], %w[1 4 7], %w[2 5 8], %w[3 6 9], %w[1 5 9], %w[3 5 7]]
