+++ lib/player.rb
@@ -1,0 +1,8 @@
+class Player
+  attr_reader :name
+  attr_accessor :choices
+  def initialize(name)
+    @name = name
+    @choices = []
+  end
+end
